fix: Saying JULIET instead of HOURS
* Updated HORUS to JULIET

As per [this Wikipedia article](https://en.wikipedia.org/wiki/24-hour_clock#Military_time) JULIET seems much more appropriate than HOURS as `J` is the letter of the English alphabet designated as the local time. I believe it is safe to assume that local time is the default time on any Pebble watch.